+++ categorical.go
@@ -9,6 +9,6 @@ import (
 // ExtractCategorical returns a vector that is a positive boolean for a
 // deterministic categorical feature ID
 func ExtractCategorical(field, category string) string {
-	featureID := murmur3.Sum32([]byte(uniqueHashPrefixStr + field + category))
+	featureID := int32(murmur3.Sum32([]byte(uniqueHashPrefixStr + field + category)))
 	return fmt.Sprintf("%d:%d", featureID, 1)
 }
