+++ text.go
@@ -16,7 +16,7 @@ func ExtractText(text, delim string) []string {
 		// this works okay for sparse technical logs
 		// otherwise you might want to use an IDF transformation
 		// to under-weight less meaningfull words/tokens
-		fID := murmur3.Sum32([]byte(v))
+		fID := int32(murmur3.Sum32([]byte(v)))
 		features = append(features, fmt.Sprintf("%d:1", fID))
 	}
